// src/types/lead.ts
export type Lead = Partial<{
  leadSource: string;
  firstName: string;
  lastName: string;
  company: string;
  country: string;
  state: string;
  phone: string;
  email: string;
  employees: string;
  useCase: string;
  product: string;
  orgType: string;
  companyIndustry: string;
  industryOfInterest: string;
  coverageRegion: string;
  revenueCriteria: string;
  comments: string;
}>;